[ACTIVITI-3516]
- rename cfg package to conf
- rename autoconfigure package to conf
