+++ activiti-engine/src/main/java/org/activiti/engine/conf/AbstractProcessEngineConfigurator.java
@@ -10,7 +10,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.engine.cfg;
+package org.activiti.engine.conf;
 
 import org.activiti.engine.impl.cfg.ProcessEngineConfigurationImpl;
