+++ activiti-engine/src/main/java/org/activiti/engine/conf/security/CommandExecutorFactory.java
@@ -10,7 +10,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.engine.cfg.security;
+package org.activiti.engine.conf.security;
 
-public interface ExecutorContext {
+import org.activiti.engine.impl.util.CommandExecutor;
+
+/**
+
+ */
+public interface CommandExecutorFactory {
+
+    CommandExecutor createExecutor(ExecutorContext executorContext);
+
 }
