+++ activiti-engine/src/main/java/org/activiti/engine/conf/security/ExecutorContext.java
@@ -10,17 +10,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.engine.cfg.security;
+package org.activiti.engine.conf.security;
 
-public class CommandExecutorContext {
-
-    private static CommandExecutorFactory shellCommandExecutorFactory;
-
-    public static void setShellExecutorContextFactory(CommandExecutorFactory shellCommandExecutorFactory) {
-        CommandExecutorContext.shellCommandExecutorFactory = shellCommandExecutorFactory;
-    }
-
-    public static CommandExecutorFactory getShellCommandExecutorFactory() {
-        return shellCommandExecutorFactory;
-    }
+public interface ExecutorContext {
 }
