+++ activiti-engine/src/main/java/org/activiti/engine/test/profiler/ActivitiProfiler.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-import org.activiti.engine.cfg.ProcessEngineConfigurator;
+import org.activiti.engine.conf.ProcessEngineConfigurator;
 import org.activiti.engine.impl.cfg.ProcessEngineConfigurationImpl;
 import org.activiti.engine.impl.interceptor.CommandInterceptor;
