+++ activiti-engine/src/test/java/org/activiti/engine/test/api/v6/AbstractActviti6Test.java
@@ -11,7 +11,7 @@ import org.activiti.engine.ProcessEngineLifecycleListener;
 import org.activiti.engine.RepositoryService;
 import org.activiti.engine.RuntimeService;
 import org.activiti.engine.TaskService;
-import org.activiti.engine.cfg.ProcessEngineConfigurator;
+import org.activiti.engine.conf.ProcessEngineConfigurator;
 import org.activiti.engine.impl.ProcessEngineImpl;
 import org.activiti.engine.impl.cfg.ProcessEngineConfigurationImpl;
 import org.activiti.engine.test.ActivitiRule;
