+++ activiti-engine/src/test/resources/org/activiti/engine/test/bpmn/usertask/TaskDueDateExtensionsTest.activiti.cfg.xml
@@ -28,7 +28,7 @@
         <property name="mailServers">
             <map>
                 <entry key="myEmailTenant">
-                    <bean class="org.activiti.engine.cfg.MailServerInfo">
+                    <bean class="org.activiti.engine.conf.MailServerInfo">
                         <property name="mailServerHost" value="localhost"/>
                         <property name="mailServerPort" value="5025"/>
                         <property name="mailServerUseSSL" value="false"/>
