+++ activiti-spring-process-extensions/src/main/java/org/activiti/spring/process/conf/ProcessExtensionsConfiguratorAutoConfiguration.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.activiti.spring.process.autoconfigure;
+package org.activiti.spring.process.conf;
 
-import org.activiti.engine.cfg.AbstractProcessEngineConfigurator;
+import org.activiti.engine.conf.AbstractProcessEngineConfigurator;
 import org.activiti.engine.impl.cfg.ProcessEngineConfigurationImpl;
 import org.activiti.spring.process.ProcessVariablesInitiator;
 import org.springframework.context.annotation.Configuration;
