+++ activiti-spring-process-extensions/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.activiti.spring.process.autoconfigure.ProcessExtensionsAutoConfiguration,\
-  org.activiti.spring.process.autoconfigure.ProcessExtensionsConfiguratorAutoConfiguration
+    org.activiti.spring.process.conf.ProcessExtensionsAutoConfiguration,\
+    org.activiti.spring.process.conf.ProcessExtensionsConfiguratorAutoConfiguration
