+++ activiti-spring-process-extensions/src/test/java/org/activiti/spring/process/ProcessVariablesInitiatorIT.java
@@ -24,8 +24,8 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.activiti.engine.ActivitiException;
 import org.activiti.engine.repository.ProcessDefinition;
-import org.activiti.spring.process.autoconfigure.ProcessExtensionsAutoConfiguration;
-import org.activiti.spring.process.autoconfigure.ProcessExtensionsConfiguratorAutoConfiguration;
+import org.activiti.spring.process.conf.ProcessExtensionsAutoConfiguration;
+import org.activiti.spring.process.conf.ProcessExtensionsConfiguratorAutoConfiguration;
 import org.activiti.spring.process.model.ProcessExtensionModel;
 import org.junit.Before;
 import org.junit.Test;
